intellitopic/requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
openai==1.3.7
httpx==0.23.3
requests
python-multipart